Small bug fix
I previously fixed this by increasing the "Wait" timer without realising that I completely missed to add another wait timer. Fixed now. The wait times may be tweaked, but some people are having issues with this closer to 0. Don't know for sure if it's the timer though, just playing it safe.